requirements.txt
pandas
numpy
scikit-learn
joblib
google-auth
google-api-python-client
